+++ trikotFoundation/src/nativeMain/kotlin/com/mirego/trikot/foundation/date/Date.kt
@@ -1,17 +1,13 @@
 package com.mirego.trikot.foundation.date
 
+import com.mirego.trikot.foundation.system.OSVersion
+import com.mirego.trikot.foundation.system.osVersionAtLeast
 import platform.Foundation.NSDate
 import platform.Foundation.NSISO8601DateFormatWithFractionalSeconds
 import platform.Foundation.NSISO8601DateFormatWithInternetDateTime
 import platform.Foundation.NSISO8601DateFormatter
-import platform.Foundation.NSNumericSearch
-import platform.Foundation.NSOrderedAscending
-import platform.Foundation.NSString
-import platform.Foundation.compare
-import platform.Foundation.create
 import platform.Foundation.dateByAddingTimeInterval
 import platform.Foundation.timeIntervalSince1970
-import platform.UIKit.UIDevice
 import kotlin.time.Duration
 import kotlin.time.ExperimentalTime
 
@@ -24,7 +20,7 @@ actual class Date(val nsDate: NSDate) {
     }
 
     actual operator fun plus(duration: Duration): Date {
-        return Date(nsDate.dateByAddingTimeInterval(((duration.toLongMilliseconds() / 1000.0))))
+        return Date(nsDate.dateByAddingTimeInterval(((duration.inWholeMilliseconds / 1000.0))))
     }
 
     actual companion object DateFactory {
@@ -38,7 +34,7 @@ actual class Date(val nsDate: NSDate) {
         @ExperimentalUnsignedTypes
         actual fun fromISO8601(isoDate: String): Date {
             return when {
-                iosVersionGreaterThan(IOSVersion.IOS_11) ->
+                osVersionAtLeast(OSVersion.iOS_11, OSVersion.macOS_10_13, OSVersion.tvOS_11, OSVersion.watchOS_4) ->
                     getFormatterWithCorrespondingFormatOptions(isoDate)
                         .dateFromString(isoDate)
                 else ->
@@ -53,7 +49,6 @@ actual class Date(val nsDate: NSDate) {
             return Date(NSDate(timeIntervalSinceReferenceDate = (epoch.toDouble() / 1000.0) - epochReferenceDateDelta))
         }
 
-        // NSISO8601DateFormatWithFractionalSeconds is iOS 11+
         @ExperimentalUnsignedTypes
         private fun getFormatterWithCorrespondingFormatOptions(date: String): NSISO8601DateFormatter =
             when {
@@ -73,30 +68,6 @@ actual class Date(val nsDate: NSDate) {
             }
 
         private fun containsFractionalSeconds(date: String): Boolean = date.lastIndexOf(".") != -1
-
-        private enum class IOSVersion {
-            IOS_10,
-            IOS_11,
-            IOS_12,
-            IOS_13,
-            IOS_14;
-
-            val stringValue: String
-                get() = when (this) {
-                    IOS_10 -> "10.0"
-                    IOS_11 -> "11.0"
-                    IOS_12 -> "12.0"
-                    IOS_13 -> "13.0"
-                    IOS_14 -> "14.0"
-                }
-        }
-
-        private fun iosVersionGreaterThan(version: IOSVersion): Boolean =
-            NSString.create(string = UIDevice.currentDevice.systemVersion)
-                .compare(
-                    string = version.stringValue,
-                    options = NSNumericSearch
-                ) != NSOrderedAscending
     }
 
     actual operator fun compareTo(other: Date): Int {