+++ trikotFoundation/src/nativeMain/kotlin/com/mirego/trikot/foundation/system/OSPlatform.kt
@@ -1,0 +1,10 @@
+package com.mirego.trikot.foundation.system
+
+enum class OSPlatform {
+    iOS,
+    tvOS,
+    watchOS,
+    macOS
+}
+
+expect var currentPlatform: OSPlatform