+++ trikotFoundation/src/nativeMain/kotlin/com/mirego/trikot/foundation/timers/PlatformTimer.kt
@@ -11,7 +11,11 @@ import kotlin.time.ExperimentalTime
 actual class PlatformTimer actual constructor(delay: Duration, repeat: Boolean, block: () -> Unit) : Timer {
     private val timerBlock: (NSTimer?) -> Unit = { _ -> block() }
     private val frozenTimerBlock = timerBlock.also { freeze(it) }
-    private val timer = NSTimer.timerWithTimeInterval((delay.toLongMilliseconds() / 1000.0), repeat, frozenTimerBlock)
+    private val timer = NSTimer.timerWithTimeInterval(
+        (delay.inWholeMilliseconds / 1000.0),
+        repeat,
+        frozenTimerBlock
+    )
 
     init {
         NSRunLoop.mainRunLoop.addTimer(timer, NSDefaultRunLoopMode)