+++ trikotFoundation/src/tvosMain/kotlin/com/mirego/trikot/foundation/system/OSPlatform.kt
@@ -1,0 +1,3 @@
+package com.mirego.trikot.foundation.system
+
+actual var currentPlatform = OSPlatform.tvOS