+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the library will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2.2.0
+> 2021-06-04
+
+- Added `watchOS` and `macOS` targets
+- Updated Kotlin version to `1.5.10`
+- Added OS version validation for all supported native targets
+
 ## 2.1.0
 > 2021-05-10